# src/main/java/Lesson09_hw/openapi.yaml
openapi: 3.0.3
info:
  title: Заказ на ресурсы облака
  version: 0.0.1
servers:
  - url: http://localhost:8080/api/v1/
    description: Dev server
paths:
  /clouds:
    get:
      summary: Метод получения списка ресурсов на облако
      tags:
        - Clouds
      operationId: getAllClouds
      responses:
        "200":
          description: Успешный ответ со списком ресурсов в облаке
          content:
            adplication/json:
              schema:
                $ref: "#/components/schemas/Resources"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        "500":
          description: Unexpected error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        "default":
          description: Всё остальное
          content:
            adplication/json:
              schema:
                $ref: "#/components/schemas/Error"
    post:
      summary: Метод создания заказа на облако
      tags:
        - Clouds
      operationId: createCloud
      requestBody:
        required: true
        content:
          adplication/json:
            schema:
              $ref: "#/components/schemas/Error"
      responses:
        "200":
          description: Успешный ответ создания заказа на облако
          content:
            adplication/json:
              schema:
                $ref: "#/components/schemas/Cloud"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        "500":
          description: Unexpected error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        "default":
          description: Всё остальное
          content:
            adplication/json:
              schema:
                $ref: "#/components/schemas/Error"
  /clouds/{cloud_id}:
    get:
      summary: Метод получения заказа на облако по ID
      tags:
        - Clouds
      operationId: getCloudById
      parameters:
        - name: cloud_id
          in: path
          required: true
          description: Идентификатор заказа облака
          schema:
            type: string
          example: f102b615
      responses:
        "200":
          description: Успешный ответ со списком ресурсов в облаке
          content:
            adplication/json:
              schema:
                $ref: "#/components/schemas/Clouds"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        "500":
          description: Unexpected error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        "default":
          description: Всё остальное
          content:
            adplication/json:
              schema:
                $ref: "#/components/schemas/Error"
    delete:
      summary: Метод отмены заказа по ID
      tags:
        - Clouds
      operationId: cancelCloudById
      parameters:
        - name: cloud_id
          in: path
          required: true
          description: Иденьтфикатор заказа облака
          schema:
            type: string
          example: f102b615
      responses:
        "200":
          description: Успешный ответ заказом облака по ID
          content:
            application/json: {}
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        "500":
          description: Unexpected error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        "default":
          description: Всё остальное
          content:
            adplication/json:
              schema:
                $ref: "#/components/schemas/Error"
    
components:
  schemas:
    Cloud:
      type: object
      required:
        - id_server
        - OS
        - RAM
      properties:
        cloud_id:
          type: string
          example: f102b615
        id_server:
          type: string
          example: f102b615
        OS:
          type: string
          enum:
            - Windows
            - Linux
          description: Операционная система сервера
        RAM:
          type: string
          example: 128
    Resources:
      type: object
      required:
        - OS
        - RAM
        - HDD
        - CPU
      properties:
        OS:
          type: string
          enum:
            - Windows
            - Linux
          description: Операционная система сервера
        RAM:
          type: string
          example: 128
        HDD:
          type: string
          example: 128
        CPU:
          type: string
          example: 128
    Clouds:
      type: array
      items:
        $ref: "#/components/schemas/Cloud"
    Error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: integer
        message:
          type: string